+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3tuning (development version)
+
 # mlr3tuning 1.6.1
 
 * compatibility: mlr3 1.7.2